requirements.txt
flask>=2.3.3
werkzeug>=2.3.7
flask-login>=0.6.2
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
sqlalchemy>=2.0.21
python-dotenv>=1.0.0
rich>=13.5.2
google-generativeai>=0.3.1
flask-cors>=4.0.0
gunicorn>=21.2.0
streamlit>=1.32.0 